Update GUI.py
I completely restructured GUI.py to be in line with object-oriented programming. I am uploading it here on my branch but still don't know how to push it. If anybody wants to do that, it would be great. I also included helpful comments regarding possible integration for other two searches.